requirements.txt
torch==2.7.0+cu128 --extra-index-url https://download.pytorch.org/whl/cu128  
torchvision
torchaudio==2.7.0+cu128   --extra-index-url https://download.pytorch.org/whl/cu128
transformers 
sentencepiece 
huggingface-hub 
accelerate 
langchain 
langchain-huggingface 
langchain-community 
faiss-cpu 
hazm 
sentence-transformers 
bitsandbytes 
jupyterlab-git==0.51.1 
rank_bm25 
protobuf 
langdetect
jupyter 
ipywidgets
llama-cpp-python
nbimporter
pytest